Add admin dashboard layout and components; remove deprecated pages and update service links
