+++ package.json
@@ -14,6 +14,8 @@
     "@tailwindcss/line-clamp": "^0.4.4",
     "@vercel/analytics": "^1.5.0",
     "antd": "^5.24.0",
+    "axios": "^1.7.9",
+    "cloudinary": "^2.5.1",
     "framer-motion": "^12.4.2",
     "mongodb": "^6.13.0",
     "mongoose": "^8.10.1",
