+++ src/app/admin/page.js
@@ -1,0 +1,7 @@
+export default function AdminDashboard() {
+  return (
+    <div>
+      <h1 className="text-2xl font-bold h-screen">Admin Dashboard</h1>
+    </div>
+  );
+}
