--- src/app/business-digitalization/dc-sdn-solution/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function DCSDNSoluntion() {
-  return <ComingSoon />;
-}
