--- src/app/business-digitalization/dc-switching-solution/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function DCSwitchingSoluntion() {
-  return <ComingSoon />;
-}
