--- src/app/business-digitalization/network-security/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function NetworkSecurity() {
-  return <ComingSoon />;
-}
