--- src/app/business-digitalization/wifi-wireless-network/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function WifiWirelessNetwork() {
-  return <ComingSoon />;
-}
