--- src/app/system-integration/dc-infra-solution/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function DCInfrastructureSolution() {
-  return <ComingSoon />;
-}
