--- src/app/system-integration/fiber-network/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function FiberNetwork() {
-  return <ComingSoon />;
-}
