--- src/app/system-integration/network-business/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function NetworkBusiness() {
-  return <ComingSoon />;
-}
