--- src/app/system-integration/server-storage/page.js
@@ -1,5 +1,0 @@
-import ComingSoon from "@/components/shared/ComingSoon";
-
-export default function ServerStorage() {
-  return <ComingSoon />;
-}
